+++ api-tests/uniqueness/unique.test.js
@@ -1,5 +1,5 @@
-const { Text } = require('@keystone-alpha/fields');
-const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystone-alpha/test-utils');
+const { Text } = require('@keystonejs/fields');
+const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystonejs/test-utils');
 const cuid = require('cuid');
 
 function setupKeystone(adapterName) {
@@ -41,7 +41,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           });
 
           expect(errors).toHaveProperty('0.message');
-          expect(errors[0].message).toEqual(expect.stringMatching(/duplicate key/));
+          expect(errors[0].message).toEqual(expect.stringMatching(/duplicate key|to be unique/));
         })
       );
 
@@ -59,7 +59,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           });
 
           expect(errors).toHaveProperty('0.message');
-          expect(errors[0].message).toEqual(expect.stringMatching(/duplicate key/));
+          expect(errors[0].message).toEqual(expect.stringMatching(/duplicate key|to be unique/));
         })
       );