+++ packages/app-admin-ui/client/classes/List.js
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import { arrayToObject, mapKeys, omit } from '@keystone-alpha/utils';
+import { arrayToObject, mapKeys, omit } from '@keystonejs/utils';
 
 export const gqlCountQueries = lists => gql`{
   ${lists.map(list => list.countQuery()).join('\n')}