+++ packages/app-admin-ui/client/components/UpdateManyItemsModal.js
@@ -6,7 +6,7 @@ import { Button, LoadingButton } from '@arch-ui/button';
 import Drawer from '@arch-ui/drawer';
 import { FieldContainer, FieldLabel, FieldInput } from '@arch-ui/fields';
 import Select from '@arch-ui/select';
-import { omit, arrayToObject, countArrays } from '@keystone-alpha/utils';
+import { omit, arrayToObject, countArrays } from '@keystonejs/utils';
 import { LoadingIndicator } from '@arch-ui/loading';
 
 import { validateFields } from '../util';