+++ packages/arch/packages/navbar/src/index.js
@@ -1,6 +1,7 @@
 // @flow
 export {
   BrandItem,
+  NavIcon,
   PrimaryNav,
   PrimaryNavItem,
   PrimaryNavHeading,