+++ website/src/components/Header.js
@@ -6,7 +6,7 @@ import { jsx } from '@emotion/core';
 import { colors, gridSize } from '@arch-ui/theme';
 
 import logosvg from '../assets/logo.svg';
-import { Container } from '../components';
+import { Container, SocialIconsNav } from '../components';
 import { media, mediaOnly, mediaMax } from '../utils/media';
 
 export const HEADER_HEIGHT = 60;
@@ -73,7 +73,7 @@ const Logo = () => (
           marginLeft: '0.5em',
         }}
       >
-        v5.x alpha
+        v5
       </span>
     </Link>
   </div>
@@ -83,7 +83,6 @@ const NavItem = ({ as, lgOnly, ...props }) => {
   return (
     <li>
       <Tag
-        href="https://github.com/keystonejs/keystone-5"
         css={{
           alignItems: 'center',
           background: 0,
@@ -133,14 +132,16 @@ const Nav = ({ toggleMenu }) => (
           {name}
         </NavItem>
       ))}
-      <NavItem
-        href="https://github.com/keystonejs/keystone-5"
-        title="Opens in new window"
-        target="_blank"
-      >
-        GitHub
-        <NewWindowIcon />
-      </NavItem>
+      <li>
+        <SocialIconsNav
+          css={{
+            marginLeft: '2rem',
+            [mediaMax.sm]: {
+              display: 'none',
+            },
+          }}
+        />
+      </li>
       {toggleMenu && (
         <NavItem
           as="button"
@@ -160,26 +161,6 @@ const Nav = ({ toggleMenu }) => (
       )}
     </List>
   </nav>
-);
-const NewWindowIcon = () => (
-  <span css={{ marginLeft: gridSize / 2, opacity: 0.6 }}>
-    <svg x="0px" y="0px" viewBox="0 0 100 100" width="15" height="15" className="css-19vhmgv">
-      <path
-        fill="currentColor"
-        d="
-      M18.8,85.1h56l0,0c2.2,0,4-1.8,4-4v-32h-8v28h-48v-48h28v-8h-32l0,
-      0c-2.2,0-4,1.8-4,4v56C14.8,83.3,16.6,85.1,18.8,85.1z
-    "
-      />
-      <polygon
-        fill="currentColor"
-        points="
-      45.7,48.7 51.3,54.3 77.2,28.5 77.2,37.2 85.2,37.2 85.2,14.9 62.8,
-      14.9 62.8,22.9 71.5,22.9
-      "
-      />
-    </svg>
-  </span>
 );
 
 // ==============================